chore: simplify itemdetails structure
